Tratamento exceções

+++ src/main/java/br/ufg/inf/backend/stp/especialidade/EspecialidadeController.java
@@ -7,6 +7,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -14,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.ufg.inf.backend.stp.ApiResponse;
-import jakarta.websocket.server.PathParam;
 
 @RestController
 @RequestMapping("/especialidade")
@@ -26,37 +26,76 @@ public class EspecialidadeController {
 	@Autowired
 	private ApiResponse<Especialidade> response;
 
+	@Autowired
+	private ApiResponse<List<Especialidade>> responseList;
+
+	@Autowired
+	ApiResponse<Void> responseVoid ;
+
 	@GetMapping
-	public ResponseEntity<List<Especialidade>> listar() {
+	public ResponseEntity<ApiResponse<List<Especialidade>>> listar() {
 		try {
 			List<Especialidade> especialidades = service.listar();
-			return ResponseEntity.ok(especialidades);
+			responseList.setData(especialidades);
+			responseList.setMessage("Especialidades listadas com sucesso");
+			responseList.setSuccess(true);
+			return ResponseEntity.ok(responseList);
 		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+			responseList.setData(null);
+			responseList.setMessage("Erro ao listar especialidades: " + e.getMessage());
+			responseList.setSuccess(false);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseList);
 		}
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<Especialidade> obter(@PathParam(value = "id") Long especialidadeId) {
+	public ResponseEntity<ApiResponse<Especialidade>> obter(@PathVariable("id") Long especialidadeId) {
+		if (especialidadeId == null) {
+			response.setMessage("ID da especialidade não pode ser nulo");
+			response.setSuccess(false);
+			return ResponseEntity.badRequest().body(response);
+		}
 		try {
 			Especialidade especialidade = service.obter(especialidadeId);
 			if (especialidade != null) {
-				return ResponseEntity.ok(especialidade);
+				response.setData(especialidade);
+				response.setMessage("Especialidade obtida com sucesso");
+				response.setSuccess(true);
+				return ResponseEntity.ok(response);
 			} else {
-				return ResponseEntity.notFound().build();
+				response.setData(null);
+				response.setMessage("Especialidade não encontrada");
+				response.setSuccess(false);
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
 			}
 		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+			response.setData(null);
+			response.setMessage("Erro ao obter especialidade: " + e.getMessage());
+			e.printStackTrace();
+			response.setSuccess(false);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
 		}
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> remover(@PathParam(value = "id") Long especialidadeId) {
+	public ResponseEntity<ApiResponse<Void>> remover(@PathVariable("id") Long especialidadeId) {
+
+		if (especialidadeId == null) {
+			responseVoid.setMessage("ID da especialidade não pode ser nulo");
+			responseVoid.setSuccess(false);
+			return ResponseEntity.badRequest().body(responseVoid);
+		}
 		try {
 			service.remover(especialidadeId);
-			return ResponseEntity.noContent().build();
+			responseVoid.setData(null);
+			responseVoid.setMessage("Especialidade removida com sucesso");
+			responseVoid.setSuccess(true);
+			return ResponseEntity.ok(responseVoid);
 		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+			responseVoid.setData(null);
+			responseVoid.setMessage("Erro ao remover especialidade: " + e.getMessage());
+			responseVoid.setSuccess(false);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(responseVoid);
 		}
 	}
 
@@ -77,16 +116,25 @@ public class EspecialidadeController {
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<Especialidade> atualizar(@PathParam(value = "id") Long especialidadeId, @RequestBody Especialidade especialidade) {
+	public ResponseEntity<ApiResponse<Especialidade>> atualizar(@PathVariable("id") Long especialidadeId, @RequestBody Especialidade especialidade) {
 		try {
 			Especialidade especialidadeAtualizada = service.salvar(especialidadeId, especialidade);
 			if (especialidadeAtualizada != null) {
-				return ResponseEntity.ok(especialidadeAtualizada);
+				response.setData(especialidadeAtualizada);
+				response.setMessage("Especialidade atualizada com sucesso");
+				response.setSuccess(true);
+				return ResponseEntity.ok(response);
 			} else {
-				return ResponseEntity.notFound().build();
+				response.setData(null);
+				response.setMessage("Especialidade não encontrada");
+				response.setSuccess(false);
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
 			}
 		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+			response.setData(null);
+			response.setMessage("Erro ao atualizar especialidade: " + e.getMessage());
+			response.setSuccess(false);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
 		}
 	}
